doc: Add example output to README
The README does not provide any example output for the random number
generation call. Unfortunately, the actual output of the call includes 4
warnings, which makes it easy to miss the succesfull generation of
a random number. An example output might improve this situation and help
the user to recognize the successful call as such.

Signed-off-by: Julien Hachenberger <julien.hachenberger@sit.fraunhofer.de>
Signed-off-by: GaelGuegan <gael.guegan@non.se.com>