Set TCTI option through an environment variable
The default build configuration should be for the common case. In the common
case, a platform has a single TPM, and may or may not have an AB/RM daemon.
General applications should only need one TCTI, and the ESYS library should be
configured to use it by default.

In some uncommon cases, the TCTI would need to be configured at run time (e.g.
developing with a simulator TPM). For these cases, the `--enable-tctienvvar`
build-time configuration flag is available to enable the `tpm2tss` Open SSL
engine to get the desired TCTI option through an environment variable.

configure.ac:
    Added configuration flag "--enable-tctienvvar". This flag sets the
    preprocessor flag, `ENABLE_TCTIENVVAR`, to enable the behaviour mentioned
    above.
tpm2-tss-engine.c:
    Modified the engine initialization function to set default TCTI option from
    an enviornment variable.

Signed-off-by: Safayet Ahmed <safayet.ahmed@ge.com>
Signed-off-by: GaelGuegan <gael.guegan@non.se.com>