Support setting binary passwords
Adds support for using passwords with non-printable characters from the command
line by using a hex encoded string (see the man-page for usage details). Also
adds the same support to the engine API, including the ability to use a non-zero
size to bypass string interpretation entirely (thus using the buffer contents
as-is regardless of its contents).

Signed-off-by: Emil Pettersson <khwaaj@gmail.com>