tpm2tss_tpm2data_readtpm(): be smarter about guessing emptyAuth.
With this patch tpm2tss_tpm2data_readtpm will assume empty Auth not only
if it has explicitely tested it for NODA objects, but also if it finds
an empty authPolicy in public areas.

With this change using tpm2_create'ed objects with default attributes
grows convenience as it does not ask for password if no auth policy is set.

Signed-off-by: Alexander Sack <asac@pantacor.com>