Use tss2-tctildr
This new library encapsulates handling of the TCTI context, allowing us
to remove our own TCTI management. It is available in tpm2-tss 2.3.

Since we do not need to manage the dlhandle manually any more, the
ESYS_AUXCONTEXT is now useless and can be replaced by a normal
ESYS_CONTEXT. This will be cleaned up in the next commit.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>