rsa: add ability to digest & sign with restricted key
The TPM architecture mandates that a restricted signing key may only
sign a digest that has been produced by the TPM. Having a TPM key be
restricted is useful for implementing attestation using the key.

The current mechanism for signing data using OpenSSL with an RSA key
is through the RSA private encrypt method. This is too low-level, since it
relies on the digest having already been done (in software). The EVP
pkey object is the right level, since OpenSSL exposes EVP_DigestSign*
functions that operate on it.

So implement methods on EVP pkey for digest and sign for the RSA type,
inheriting other EVP pkey methods from the built-in RSA type in
OpenSSL. The digest and sign operation is implemented by making use of
the TPM to perform the digest and then passing the validation ticket,
which is required for a restricted key, (along with digest data) into
the subsequent sign operation in the TPM.

Signed-off-by: Rob Shearman <rob@graphiant.com>